Makefile: Unix build
This is fairly trivial, but both saves developers a bit of work to
figure out how to build this by and and lets Linux users know before
they download this that it should build and run on their system.
README updates mentioning the build configuration files are included.

Tested on Debian 9.